HW-3 MarketPresenter 생성 & 기존 프로젝트에 MVP 패턴 적용
