+++ preludezdev/ArchitectureStudy/app/src/main/java/com/example/architecturestudy/ui/MarketFragment.kt
@@ -1,29 +1,30 @@
 package com.example.architecturestudy
 
-import android.graphics.Color
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.architecturestudy.data.Coin
-import com.example.architecturestudy.data.CoinTickerResponse
 import com.example.architecturestudy.data.source.CoinsRepository
 import com.example.architecturestudy.data.source.local.CoinsLocalDataSource
 import com.example.architecturestudy.data.source.remote.CoinsRemoteDataSource
 import com.example.architecturestudy.network.RetrofitHelper
+import com.example.architecturestudy.ui.MarketContract
+import com.example.architecturestudy.ui.MarketPresenter
 import com.example.architecturestudy.ui.RecyclerViewAdapter
-import com.example.architecturestudy.util.Util
 import kotlinx.android.synthetic.main.fragment_market.*
 
-class MarketFragment : Fragment() {
+class MarketFragment : Fragment(), MarketContract.View {
+
+    private lateinit var presenter: MarketContract.Presenter // Presenter 프로퍼티 선언
 
     private val repository = CoinsRepository.getInstance(
         CoinsRemoteDataSource.getInstance(RetrofitHelper.getInstance().coinApiService),
         CoinsLocalDataSource.getInstance()
     )
+
     private val rvAdapter = RecyclerViewAdapter()
 
     companion object {
@@ -57,73 +58,16 @@ class MarketFragment : Fragment() {
 
         var keyMarket = arguments?.get("KEY_MARKET") as String
 
-        loadData(keyMarket) // 해당 마켓의 데이터 불러오기
+        presenter = MarketPresenter(this, repository) // Presenter 객체 생성해서 할당
+        presenter.loadData(keyMarket) // Presenter 를 통해 해당 마켓의 데이터 불러오기
     }
 
-    private fun loadData(keyMarket: String) {
+    override fun setTickerData(data: List<Coin>) {
+        rvAdapter.setData(data)
+    }
+
+    override fun clearTickerData() {
         rvAdapter.clearData()
-
-        repository
-            .getAllMarket({ coinMarketResponse ->
-
-                if (coinMarketResponse != null) {
-
-                    val targetTickers = coinMarketResponse.filter {
-                        it.market.split('-')[0] == keyMarket
-                    }.joinToString(separator = ",") { it.market }
-
-                    repository
-                        .getCoinTickers(targetTickers, { coinTickerResponse ->
-                            //map() 스트림 함수 : 컬렉션 내 인자를 변환하여 반환
-                            if (coinTickerResponse != null) {
-                                rvAdapter.setData(coinTickerResponse.map(::convertTickerIntoCoin))
-                            }
-                        }, { onFailCallback(it) })
-                }
-            }, { onFailCallback(it) })
-    }
-
-    private fun onFailCallback(errorMsg: String) {
-        Log.d("test", errorMsg)
-    }
-
-    private fun convertTickerIntoCoin(ticker: CoinTickerResponse): Coin {
-        //코인 이름
-        var market = ticker.market.split("-")[1]
-
-        //현재가
-        var tradePrice = when {
-            ticker.tradePrice > 1_000 ->
-                Util.convertBigNumberToStdString(ticker.tradePrice.toInt())
-            ticker.tradePrice > 2 ->
-                String.format("%.2f", ticker.tradePrice)
-            else ->
-                String.format("%.8f", ticker.tradePrice)
-        }
-
-        //전일대비
-        var signedChangeRate = String.format("%.2f", ticker.signedChangeRate * 100) + "%"
-
-        //전일대비 색깔 지정
-        val coinColor = if (signedChangeRate.startsWith('-')) Color.BLUE else Color.RED
-
-        //거래대금
-        var accTradePriceH = when {
-            ticker.accTradePriceH > 10_000_000 -> {
-                Util.convertBigNumberToStdString((ticker.accTradePriceH / 1000000).toInt()) + "M"
-            }
-
-            ticker.accTradePriceH > 100_000 ->
-                Util.convertBigNumberToStdString(ticker.accTradePriceH.toInt() / 1000) + "k"
-
-            ticker.accTradePriceH > 1_000 ->
-                Util.convertBigNumberToStdString(ticker.accTradePriceH.toInt())
-
-            else ->
-                String.format("%.3f", ticker.accTradePriceH)
-        }
-
-        return Coin(market, tradePrice, signedChangeRate, accTradePriceH, coinColor)
     }
 
 }
